[REFACT] Point 도메인에 대한 유효성 검사 로직 이동 및 수정
