+++ src/main/java/com/programmers/smrtstore/domain/point/application/dto/req/PointRequest.java
@@ -1,9 +1,7 @@
 package com.programmers.smrtstore.domain.point.application.dto.req;
 
-import com.programmers.smrtstore.core.properties.ErrorCode;
 import com.programmers.smrtstore.domain.point.domain.entity.Point;
 import com.programmers.smrtstore.domain.point.domain.entity.enums.PointStatus;
-import com.programmers.smrtstore.domain.point.exception.PointException;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -20,7 +18,6 @@ public class PointRequest {
     private Long orderId;
 
     public Point toEntity(PointStatus pointStatus, Integer pointValue, boolean membershipApplyYn) {
-        validatePointValue(pointStatus, pointValue);
         return Point.builder()
             .userId(userId)
             .orderId(orderId)
@@ -30,15 +27,5 @@ public class PointRequest {
             .build();
     }
 
-    private static void validatePointValue(PointStatus pointStatus, Integer pointValue) {
-        if (pointStatus.equals(PointStatus.ACCUMULATED) || pointStatus.equals(PointStatus.USE_CANCELED)) {
-            if (pointValue < 0) {
-                throw new PointException(ErrorCode.POINT_ILLEGAL_ARGUMENT, String.valueOf(pointValue));
-            }
-        } else {
-            if (pointValue > 0) {
-                throw new PointException(ErrorCode.POINT_ILLEGAL_ARGUMENT, String.valueOf(pointValue));
-            }
-        }
-    }
+
 }
